Conflicts:
	kbe/src/server/cellapp/entity.cpp
	kbe/src/server/cellapp/witness.cpp

+++ kbe/src/lib/client_lib/clientobjectbase.cpp
@@ -1961,7 +1961,7 @@ PyObject* ClientObjectBase::__py_GetSpaceData(PyObject* self, PyObject* args)
 //-------------------------------------------------------------------------------------
 PyObject* ClientObjectBase::__py_getWatcher(PyObject* self, PyObject* args)
 {
-	int argCount = PyTuple_Size(args);
+	int argCount = (int)PyTuple_Size(args);
 	if(argCount != 1)
 	{
 		PyErr_Format(PyExc_TypeError, "KBEngine::getWatcher(): args[strpath] is error!");
@@ -2100,7 +2100,7 @@ PyObject* ClientObjectBase::__py_getWatcher(PyObject* self, PyObject* args)
 //-------------------------------------------------------------------------------------
 PyObject* ClientObjectBase::__py_getWatcherDir(PyObject* self, PyObject* args)
 {
-	int argCount = PyTuple_Size(args);
+	int argCount = (int)PyTuple_Size(args);
 	if(argCount != 1)
 	{
 		PyErr_Format(PyExc_TypeError, "KBEngine::getWatcherDir(): args[strpath] is error!");
