+++ kbe/src/lib/pyscript/vector4.cpp
@@ -248,7 +248,7 @@ PyObject* ScriptVector4::seq_slice(PyObject* self, Py_ssize_t startIndex, Py_ssi
 	Vector4& my_v = sv->getVector();
 	PyObject* pyResult = NULL;
 
-	int length = endIndex - startIndex;
+	int length = (int)(endIndex - startIndex);
 
 	if (length == VECTOR_SIZE)
 	{
@@ -269,7 +269,7 @@ PyObject* ScriptVector4::seq_slice(PyObject* self, Py_ssize_t startIndex, Py_ssi
 			{
 				Vector2 v;
 				
-				for(int i = startIndex; i < endIndex; ++i){
+				for (int i = (int)startIndex; i < (int)endIndex; ++i){
 					v[i - static_cast<int>(startIndex)] = my_v[i];
 				}
 
@@ -279,7 +279,7 @@ PyObject* ScriptVector4::seq_slice(PyObject* self, Py_ssize_t startIndex, Py_ssi
 			case 3:
 			{
 				Vector3 v;
-				for (int i = startIndex; i < endIndex; ++i){
+				for (int i = (int)startIndex; i < (int)endIndex; ++i){
 					v[i - static_cast<int>(startIndex)] = my_v[i];
 				}
 
@@ -778,7 +778,7 @@ PyObject* ScriptVector4::__py_pySet(PyObject* self, PyObject* args)
 	Vector4 v;
 
 	// 如果参数只有1个元素
-	int tupleSize = PyTuple_Size(args);
+	int tupleSize = (int)PyTuple_Size(args);
 
 	if(tupleSize == 1)
 	{
